Remove ASSIGN-SETF-MACRO's secondary test for fboundness of SETF fun.
The new code is behaviorally different in a subtle way.  The old didn't
didn't warn if a setf function was defined earlier in the same file
unless in eval-when :compile-toplevel. Whereas if (SETF F) had come
from a previously loaded file, you did get a style-warning upon
definition of the macro. Now either way style-warns.

There is no change in the signaling of a full warning if the compiler
has *assumed* a setf function to exist when the macroish setter is
defined, because that suggests something was compiled wrong.

This is almost but not entirely like the patch I proposed on sbcl-devel.
The unanticipated problems were mainly during xc because :where-from
changes what it reports for a function name that has no binding.
This might possibly warn about a few things in xc that it didn't before.