De-obfuscate DEFINE-INSTRUCTION-FORMAT.
Previously if two unrelated formats had same-named args in the same positions,
specifying {prefilter,labeler,printer} as lambda expressions could clobber
one or the other format, e.g.
  (define-instruction-format (foo 8)
    (bits :field (byte 2 2) :printer (lambda (a b c) (something a b c))))
  (define-instruction-format (bar 8)
    (bits :field (byte 2 2) :printer (lambda (a b c) (otherthing a b c))))
produced two DEFPARAMETERs for the same symbol:
 (DEFPARAMETER *BITS-0-PRINTER-WRAPPER* (LAMBDA (A B C) ...))
These are now named not to collide in most cases.

Incompatible changes- :INCLUDE is parsed by the macroexpander and not
evaluated. Access functions for bits of a dchunk are generated only if
requested via a :READER option, and they appear in the target core.

Also fixed DEFINE-ARG-TYPE not to accept a :VALUE, on account of that
not being propagated by SET-ARG-FROM-TYPE, which seemed correct;
and DEFTYPE FILTERED-VALUE-INDEX was off by 1.