Absorb pcl wrapper into SB-KERNEL:LAYOUT and freeze layout's layout.
The shortens the test for whether something is a layout,
which happens all over the place.

Numerous comments suggest that the 'for-std-class' flag could be
removed by merely checking if effective-slot-definition objects
had been placed into the layout, but I don't think it's that simple.
Slot presence as the defining factor would make SB-PCL::PCL-INSTANCE-P
return NIL for slotless standard-objects, and that's (probably) wrong.