Delete incorrect test attributable to lp#770184.
All this test proved is that internal inconsistencies exist.
`(,@X) means (APPEND X) regardless of X's value, even if an atom.
`(,@1) is 1, and likewise `(1 ,@2) is (APPEND '(1) 2) => (1 . 2).
Since (LET ((X 1)) `(,@X)) and `(,@'1) both produce 1,
suspicion ought to have been cast on this test case.

The crash was indeed a problem, with a proper fix pending.