compiler/generic/genesis: ARM wants boxed FDEFN raw-addr values.
  * Like SPARC, the ARM processors have the ability to apply an
arbitrary displacement in the jump instruction.  In the ARM case,
it is actually literally ANY ALU operation with the program
counter as a destination.  We could shift or rotate some register
left or right, add the result to a second register, and load the
result into the program counter in a single instruction.  We're
going to take advantage of this in our call sequence, but this
means fixing up the COLD-FSET handling in genesis to match the
SPARC version.