compiler/arm/vm: Swap OCFP and R7.
  * In order to generate a SIGTRAP from user-mode, we need to use
a syscall.  The OABI syscall interface stored the syscall number
in the SWI instruction as an immediate constant, but nobody uses
OABI anymore.  EABI has the syscall number as R7, but the syscall
that we need is an unboxed value, so we need that R7 be an unboxed
(non-descriptor) register.

  * We can't use LIP because the GC can rewrite it, to our
detriment.  We can't use NFP because there's a good chance that it
gets set as a component-live-tn.  We can't use NARGS because it's
used in the calling side of our calling convention and we need to
be able to generate a SIGTRAP from undefined_tramp.

  * This leaves OCFP as the only viable candidate, since we pass
the OCFP in its save location rather than the register.  Function
return uses OCFP for its original purpose, but we're less worried
about needing to SIGTRAP there.