Basic support for FIXUPs on ARM.
  * ARM can only handle 8 bits of immediate at a time, which gives
some not-particularly-nice instruction sequences.  If we must, we
can use such sequences, but the more commonly done thing is to use
a PC-relative data load in order to read up a full word at a time.

  * In order to support said typical ARM constant load scenario
for fixups, teach the WORD pseudo-instruction about fixups, and
teach genesis about ABSOLUTE fixups for ARM.