compiler/arm/{call,vm}: Add support for "implicit" value cell references.
  * There's a bit of magic behind the scenes for this, but the
compiler can occasionally decide that there's no need to create a
full "value cell" for some aspects of a closure, and instead it
takes note of the stack frame that contains the variable being
closed over and uses the ANCESTOR-FRAME-{REF,SET} VOPs to access
the variable from other frames.  Simple enough to implement.

  * And the wrinkle that I hadn't expected is that there is a hook
to allow this to be done for unboxed variables, presumably using a
separate set of VOPs, and that hook needs at least a minimal
implementation, fortunately provided for most backends.

  * This is all cribbed verbatim from the PPC backend.

  * And re-enabled the relevant parts of IR2TRAN, as usual.